modifiche al GIORNO 10

+++ GIORNO 010-Functions with outputs/5-Calculator 2.py
@@ -1,6 +1,4 @@
 
-
-
 #addizione
 def add(n1, n2):
     return n1 + n2
@@ -26,7 +24,7 @@ operations = {
 
 def calculator():
     
-    num1 = int(input("What's the first number?:\n"))
+    num1 = float(input("What's the first number?:\n"))
 
     for key in operations:
         print(key)
@@ -36,7 +34,7 @@ def calculator():
     while should_continue:
 
         operation_symbol = input("Pick an operation: ")
-        num2 = int(input("What's the next number?:\n"))
+        num2 = float(input("What's the next number?:\n"))
         calculation_function = operations[operation_symbol]
         answer = calculation_function(num1,num2)
 
@@ -57,7 +55,7 @@ def calculator():
         else: 
             should_continue = False
             calculator()
-
+            
 calculator()
